delete polls properly

+++ src/services/polls.service.js
@@ -8,7 +8,7 @@ const pollService = {
     updatePoll: (pollId, formData) => service.put(`/api/polls/${pollId}`, formData),
     vote: (pollId, optionId, userId) => service.post(`/api/polls/${pollId}/vote`, { optionId, userId }),
     closePoll: (pollId, userId) => service.put(`/api/polls/${pollId}/close`, { userId }),
-    deletePoll: (pollId) => service.delete(`/api/polls/${pollId}`),
+    deletePoll: (pollId, userId) => service.delete(`/api/polls/${pollId}/${userId}`)
 }
 
 export default pollService;
